Move logger http functions to separate package
