+++ pkg/logging/httpaccess/http_access.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package logging
+package httpaccess
 
 import (
 	"bufio"
@@ -11,11 +11,13 @@ import (
 	"time"
 
 	"github.com/sirupsen/logrus"
+
+	"github.com/ethersphere/bee/pkg/logging"
 )
 
 // NewHTTPAccessLogHandler creates a handler that will log a message after a
 // request has been served.
-func NewHTTPAccessLogHandler(logger Logger, level logrus.Level, message string) func(h http.Handler) http.Handler {
+func NewHTTPAccessLogHandler(logger logging.Logger, level logrus.Level, message string) func(h http.Handler) http.Handler {
 	return func(h http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			startTime := time.Now()
